LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-hi.html

<!DOCTYPE html><!-- webkit-test-runner [ language=hi ] -->
<html>
<head>
<script src="../../../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "hi");
</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
</html>
