LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-hi-expected.txt

PASS navigator.language is "hi"
PASS successfullyParsed is true

TEST COMPLETE
